+++ The Noob Bot/nManager/Wow/Bot/States/FightHostileTargetDamageDealerOnly.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using nManager.FiniteStateMachine;
 using nManager.Helpful;
 using nManager.Wow.Bot.Tasks;
+using nManager.Wow.Class;
 using nManager.Wow.Enums;
 using nManager.Wow.Helpers;
 using nManager.Wow.ObjectManager;
@@ -53,7 +54,7 @@ namespace nManager.Wow.Bot.States
                 Logging.Write("Please dismount as soon as is possible ! This product is passive when you are not yet in combat.");
             }
             Logging.Write("Currently attacking " + _unit.Name + " (lvl " + _unit.Level + ")");
-            ulong unkillableMob = Fight.StartFightDamageDealer(_unit.Guid);
+            Int128 unkillableMob = Fight.StartFightDamageDealer(_unit.Guid);
             if (!_unit.IsDead && unkillableMob != 0)
             {
                 Logging.Write("Can't reach " + _unit.Name + ", blacklisting it.");
@@ -86,7 +87,7 @@ namespace nManager.Wow.Bot.States
             if (Party.IsInGroup())
             {
                 List<WoWUnit> targets = new List<WoWUnit>();
-                foreach (ulong playerGuid in Party.GetPartyPlayersGUID())
+                foreach (Int128 playerGuid in Party.GetPartyPlayersGUID())
                 {
                     if (playerGuid != ObjectManager.ObjectManager.Me.Guid)
                     {
