+++ The Noob Bot/nManager/Wow/Bot/States/Grinding.cs
@@ -2,6 +2,7 @@
 using System.Threading;
 using nManager.FiniteStateMachine;
 using nManager.Helpful;
+using nManager.Wow.Class;
 using nManager.Wow.Helpers;
 using nManager.Wow.ObjectManager;
 
@@ -80,7 +81,7 @@ namespace nManager.Wow.Bot.States
         public override void Run()
         {
             Logging.Write("Player Attack " + _unit.Name + " (lvl " + _unit.Level + ")");
-            ulong unkillableMob = Fight.StartFight(_unit.Guid);
+            Int128 unkillableMob = Fight.StartFight(_unit.Guid);
             if (!_unit.IsDead && unkillableMob != 0)
             {
                 Logging.Write("Can't reach " + _unit.Name + ", blacklisting it.");
