+++ The Noob Bot/nManager/Wow/Bot/Tasks/FarmingTask.cs
@@ -13,7 +13,7 @@ namespace nManager.Wow.Bot.Tasks
 {
     public static class FarmingTask
     {
-        private static ulong _lastnode;
+        private static Int128 _lastnode;
 
         public static void Pulse(IEnumerable<WoWGameObject> nodes)
         {
